Update: formatting main.dart
